+++ persistence-module/src/main/java/it/aldinucci/todoapp/adapter/out/persistence/DeleteResetPasswordTokenJPA.java
@@ -1,6 +1,7 @@
 package it.aldinucci.todoapp.adapter.out.persistence;
 
 import java.util.Optional;
+import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Component;
 import it.aldinucci.todoapp.adapter.out.persistence.entity.ResetPasswordTokenJPA;
 import it.aldinucci.todoapp.adapter.out.persistence.repository.ResetPasswordTokenJPARepository;
 import it.aldinucci.todoapp.application.port.out.DeleteRestPasswordTokenDriverPort;
+import it.aldinucci.todoapp.util.ValidationUtils;
 
 @Component
 public class DeleteResetPasswordTokenJPA implements DeleteRestPasswordTokenDriverPort{
@@ -23,9 +25,11 @@ public class DeleteResetPasswordTokenJPA implements DeleteRestPasswordTokenDrive
 
 	@Override
 	public void delete(String token) {
-		Optional<ResetPasswordTokenJPA> optional = tokenRepo.findByToken(token);
-		if(optional.isPresent())
-			tokenRepo.delete(optional.get());
+		if (ValidationUtils.isValidUUID(token)) {
+			Optional<ResetPasswordTokenJPA> optional = tokenRepo.findByToken(UUID.fromString(token));
+			if(optional.isPresent())
+				tokenRepo.delete(optional.get());
+		}
 	}
 
 }
