+++ persistence-module/src/main/java/it/aldinucci/todoapp/adapter/out/persistence/DeleteVerificationTokenJPA.java
@@ -1,6 +1,7 @@
 package it.aldinucci.todoapp.adapter.out.persistence;
 
 import java.util.Optional;
+import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Component;
 import it.aldinucci.todoapp.adapter.out.persistence.entity.VerificationTokenJPA;
 import it.aldinucci.todoapp.adapter.out.persistence.repository.VerificationTokenJPARepository;
 import it.aldinucci.todoapp.application.port.out.DeleteVerificationTokenDriverPort;
+import it.aldinucci.todoapp.util.ValidationUtils;
 
 @Component
 public class DeleteVerificationTokenJPA implements DeleteVerificationTokenDriverPort{
@@ -22,9 +24,11 @@ public class DeleteVerificationTokenJPA implements DeleteVerificationTokenDriver
 
 	@Override
 	public void delete(String tokenCode){
-		Optional<VerificationTokenJPA> optionalToken  = tokenRepo.findByToken(tokenCode);
-		if(optionalToken.isPresent()) {
-			tokenRepo.delete(optionalToken.get());
+		if (ValidationUtils.isValidUUID(tokenCode)) {
+			Optional<VerificationTokenJPA> optionalToken  = tokenRepo.findByToken(UUID.fromString(tokenCode));
+			if(optionalToken.isPresent()) {
+				tokenRepo.delete(optionalToken.get());
+			}
 		}
 	}
